Task #6600: NenuFar classes fully tested. Beamctl ready. BeamServer needs last update.

+++ branches/LOFAR-NenuFar-Task6600/MAC/APL/PAC/ITRFBeamServer/src/NenuFarIO.h
@@ -72,7 +72,7 @@ private:
 	int		itsNCUportnr;		// portnumber the NCU system listens on
 	double	itsReconInterval;	// how often to try to reconnect
 
-	int		itsIOtimeout;		// how long to wait for answers for the NCU
+	double	itsIOtimeout;		// how long to wait for answers for the NCU (sec)
 
 	NenuFarAdmin*			itsBeams;
 	NenuFarAdmin::BeamInfo	itsCurrentBeam;
